`git ls-remote` dotfiles/homebrew-brewfile checks.
Use `git ls-remote` instead of the GitHub API for the dotfiles and
homebrew-brewfile repository checks. This means a GitHub API token is
not needed and instead it relies on the authentication that would be
used on for cloning the repository.

Additionally, try `script/bootstrap` if `script/setup` doesn't exist and
still clone the repository to the right location even if neither do.