Fix auth for OmniAuth 2.0
The omniauth gem was upgraded to 2.0 despite a breaking change: GET
/github/auth no longer works. That means that the currently-deployed
macos-strap.herokuapp.com doesn't work, either.

The fix is to switch to POST /github/auth, which this changeset does. In
order to only ever POST for auth, the flow had to change a little bit:

1. Download button submits a form to POST for auth
2. Auth callback redirects to GET strap.sh

It was previously reversed, where the button went to strap.sh, and a GET
redirect was issued if there was no auth.